+++ test/src/unit-cppapi-deletes.cc
@@ -38,7 +38,7 @@
 #include "tiledb/common/stdx_string.h"
 #include "tiledb/sm/array/array_directory.h"
 #include "tiledb/sm/c_api/tiledb_struct_def.h"
-#include "tiledb/sm/cpp_api/group_experimental.h"
+#include "tiledb/sm/cpp_api/group.h"
 #include "tiledb/sm/cpp_api/tiledb"
 
 #ifdef _WIN32
