+++ test/src/unit-cppapi-group.cc
@@ -43,7 +43,6 @@
 
 #include "tiledb/sm/c_api/tiledb.h"
 #include "tiledb/sm/cpp_api/group.h"
-#include "tiledb/sm/cpp_api/group_experimental.h"
 #include "tiledb/sm/enums/encryption_type.h"
 #include "tiledb/sm/global_state/unit_test_config.h"
 #include "tiledb/sm/misc/utils.h"
