+++ tiledb/sm/cpp_api/tiledb_experimental
@@ -41,7 +41,6 @@
 #include "dimension_label_experimental.h"
 #include "enumeration_experimental.h"
 #include "consolidation_plan_experimental.h"
-#include "group_experimental.h"
 #include "query_condition_experimental.h"
 #include "query_experimental.h"
 #include "subarray_experimental.h"
